+++ CHANGELOG.md
@@ -5,7 +5,8 @@
 
 ## [HEAD][] / unreleased
 
-* No significant changes.
+* Add faraday 2 support.
+* Extract an initial internal cURL formatter
 
 ## [2.4.2][] / 2021-11-09